# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crypto Intelligence Platform",
    "description": "Cryptocurrency Market Intelligence & Quantitative Workstation v3.0. Factor Scanner, Robust Backtesting, Real-time Streaming, and Paper Trading. Non-custodial research platform. No live execution.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/research/attribution": {
      "get": {
        "tags": [
          "Research & Validity"
        ],
        "summary": "Get forward-return factor attribution and decile report",
        "description": "Evaluate point-in-time forward returns, Information Coefficient (Pearson & Spearman), and score deciles.",
        "operationId": "get_forward_return_attribution_api_v1_research_attribution_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Forward horizon: '1d', '7d', or '30d'",
              "default": "1d",
              "title": "Horizon"
            },
            "description": "Forward horizon: '1d', '7d', or '30d'"
          },
          {
            "name": "synthetic",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Test on synthetic dataset with known ground-truth IC",
              "default": false,
              "title": "Synthetic"
            },
            "description": "Test on synthetic dataset with known ground-truth IC"
          },
          {
            "name": "min_samples",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 3,
              "description": "Minimum sample size threshold",
              "default": 30,
              "title": "Min Samples"
            },
            "description": "Minimum sample size threshold"
          },
          {
            "name": "target_ic",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 0.9,
              "minimum": -0.9,
              "description": "Target IC for synthetic test mode",
              "default": 0.18,
              "title": "Target Ic"
            },
            "description": "Target IC for synthetic test mode"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributionReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/ic": {
      "get": {
        "tags": [
          "Research & Validity"
        ],
        "summary": "Get rolling Information Coefficient time-series",
        "description": "Return chronological time series of Information Coefficients.",
        "operationId": "get_rolling_ic_series_api_v1_research_ic_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Horizon '1d', '7d', '30d'",
              "default": "1d",
              "title": "Horizon"
            },
            "description": "Horizon '1d', '7d', '30d'"
          },
          {
            "name": "synthetic",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Generate synthetic IC timeline",
              "default": false,
              "title": "Synthetic"
            },
            "description": "Generate synthetic IC timeline"
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollingICResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/trials": {
      "get": {
        "tags": [
          "Research & Validity"
        ],
        "summary": "Get counter and distribution of strategy parameter trials",
        "description": "Retrieve multiple testing trial tracking statistics for Deflated Sharpe Ratio computation.",
        "operationId": "get_strategy_trials_api_v1_research_trials_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrialsSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/model-gate": {
      "post": {
        "tags": [
          "Research & Validity"
        ],
        "summary": "Evaluate experimental ML model against production gate",
        "description": "Verify that an experimental ML model strictly beats baseline out-of-sample before production ranking.",
        "operationId": "evaluate_ml_model_gate_api_v1_research_model_gate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelGateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductionGateResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AttributionReport": {
        "properties": {
          "horizon": {
            "type": "string",
            "title": "Horizon",
            "description": "Forward return horizon ('1d', '7d', '30d')"
          },
          "sample_size": {
            "type": "integer",
            "title": "Sample Size",
            "description": "Number of snapshot observation pairs analyzed"
          },
          "data_mode": {
            "type": "string",
            "title": "Data Mode",
            "description": "Data provenance mode",
            "default": "HISTORICAL"
          },
          "pearson_ic": {
            "type": "number",
            "title": "Pearson Ic",
            "description": "Linear correlation between score and forward return"
          },
          "pearson_p_value": {
            "type": "number",
            "title": "Pearson P Value",
            "description": "P-value for H0: Pearson IC == 0"
          },
          "pearson_ci_lower": {
            "type": "number",
            "title": "Pearson Ci Lower",
            "description": "95% Bootstrap CI lower bound for Pearson IC"
          },
          "pearson_ci_upper": {
            "type": "number",
            "title": "Pearson Ci Upper",
            "description": "95% Bootstrap CI upper bound for Pearson IC"
          },
          "spearman_rank_ic": {
            "type": "number",
            "title": "Spearman Rank Ic",
            "description": "Rank correlation between score rank and return rank"
          },
          "spearman_p_value": {
            "type": "number",
            "title": "Spearman P Value",
            "description": "P-value for H0: Spearman Rank IC == 0"
          },
          "spearman_ci_lower": {
            "type": "number",
            "title": "Spearman Ci Lower",
            "description": "95% Bootstrap CI lower bound for Spearman IC"
          },
          "spearman_ci_upper": {
            "type": "number",
            "title": "Spearman Ci Upper",
            "description": "95% Bootstrap CI upper bound for Spearman IC"
          },
          "hit_rate_pct": {
            "type": "number",
            "title": "Hit Rate Pct",
            "description": "Percentage of scores > 50 yielding positive forward return"
          },
          "hit_rate_ci_lower": {
            "type": "number",
            "title": "Hit Rate Ci Lower",
            "description": "95% Bootstrap CI lower bound for hit rate"
          },
          "hit_rate_ci_upper": {
            "type": "number",
            "title": "Hit Rate Ci Upper",
            "description": "95% Bootstrap CI upper bound for hit rate"
          },
          "deciles": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Deciles",
            "description": "Performance across deciles 1-10"
          },
          "monotonicity_spread_pct": {
            "type": "number",
            "title": "Monotonicity Spread Pct",
            "description": "Decile 10 mean return minus Decile 1 mean return",
            "default": 0.0
          },
          "small_sample_warning": {
            "type": "boolean",
            "title": "Small Sample Warning",
            "description": "True if sample size N < 30 indicates high estimator variance",
            "default": false
          },
          "small_sample_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Small Sample Message",
            "description": "Plain-language warning if sample size is insufficient"
          },
          "plain_language_summary": {
            "type": "string",
            "title": "Plain Language Summary",
            "description": "Plain-language interpretation of statistical factor efficacy",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "horizon",
          "sample_size",
          "pearson_ic",
          "pearson_p_value",
          "pearson_ci_lower",
          "pearson_ci_upper",
          "spearman_rank_ic",
          "spearman_p_value",
          "spearman_ci_lower",
          "spearman_ci_upper",
          "hit_rate_pct",
          "hit_rate_ci_lower",
          "hit_rate_ci_upper"
        ],
        "title": "AttributionReport",
        "description": "Complete institutional forward-return factor attribution report."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelGateRequest": {
        "properties": {
          "model_predictions": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Model Predictions"
          },
          "baseline_predictions": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Baseline Predictions"
          },
          "actual_returns": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Actual Returns"
          },
          "min_ic_improvement": {
            "type": "number",
            "title": "Min Ic Improvement",
            "default": 0.02
          }
        },
        "type": "object",
        "required": [
          "model_predictions",
          "baseline_predictions",
          "actual_returns"
        ],
        "title": "ModelGateRequest",
        "description": "Evaluation request checking if experimental ML model passes production gate."
      },
...
      "ProductionGateResult": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "model_rank_ic": {
            "type": "number",
            "title": "Model Rank Ic"
          },
          "baseline_rank_ic": {
            "type": "number",
            "title": "Baseline Rank Ic"
          },
          "ic_delta": {
            "type": "number",
            "title": "Ic Delta"
          },
          "p_value": {
            "type": "number",
            "title": "P Value"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "evaluated_at": {
            "type": "string",
            "title": "Evaluated At"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "model_rank_ic",
          "baseline_rank_ic",
          "ic_delta",
          "p_value",
          "message",
          "evaluated_at"
        ],
        "title": "ProductionGateResult"
      },
...
      "RollingICPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "pearson_ic": {
            "type": "number",
            "title": "Pearson Ic"
          },
          "spearman_ic": {
            "type": "number",
            "title": "Spearman Ic"
          },
          "sample_size": {
            "type": "integer",
            "title": "Sample Size"
          }
        },
        "type": "object",
        "required": [
          "date",
          "pearson_ic",
          "spearman_ic",
          "sample_size"
        ],
        "title": "RollingICPoint",
        "description": "Point-in-time Information Coefficient measurement."
      },
      "RollingICResponse": {
        "properties": {
          "horizon": {
            "type": "string",
            "title": "Horizon"
          },
          "data_mode": {
            "type": "string",
            "title": "Data Mode"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/RollingICPoint"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "horizon",
          "data_mode",
          "points"
        ],
        "title": "RollingICResponse",
        "description": "Time-series of Information Coefficient observations."
      },
...
      "TrialsSummaryResponse": {
        "properties": {
          "total_trials": {
            "type": "integer",
            "title": "Total Trials"
          },
          "variance_trials": {
            "type": "number",
            "title": "Variance Trials"
          },
          "mean_sharpe": {
            "type": "number",
            "title": "Mean Sharpe"
          },
          "expected_max_sharpe": {
            "type": "number",
            "title": "Expected Max Sharpe"
          }
        },
        "type": "object",
        "required": [
          "total_trials",
          "variance_trials",
          "mean_sharpe",
          "expected_max_sharpe"
        ],
        "title": "TrialsSummaryResponse",
        "description": "Summary of strategy variants evaluated on the platform for multiple testing correction."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}